requirements.txt
Flask==3.1.2
python-dotenv==1.1.0
requests==2.32.5
PyJWT==2.10.1
